Consolidated V2ApiOperatorLocalOnlyTest into V2ApiOperatorTest
