+++ src/test/java/helper/EnabledCondition.java
@@ -2,10 +2,10 @@ package helper;
 
 import app.common.EnvUtil;
 
-public final class DisableCondition {
+public final class EnabledCondition {
     private static final String ENV = EnvUtil.getEnv("UID2_E2E_ENV");
 
-    private DisableCondition() {
+    private EnabledCondition() {
     }
 
     public static boolean isLocal() {
