--- src/test/java/suite/E2EPipelinePrivateOperatorTestSuite.java
@@ -1,16 +1,0 @@
-package suite;
-
-import org.junit.platform.suite.api.SelectClasses;
-import org.junit.platform.suite.api.Suite;
-import suite.basic.BasicTest;
-import suite.operator.V2ApiOperatorLocalOnlyTest;
-import suite.operator.V2ApiOperatorTest;
-
-@Suite
-@SelectClasses({
-        BasicTest.class,
-        V2ApiOperatorTest.class,
-        V2ApiOperatorLocalOnlyTest.class
-})
-public class E2EPipelinePrivateOperatorTestSuite {
-}
