+++ src/test/java/suite/E2EPrivateOperatorTestSuite.java
@@ -3,7 +3,6 @@ package suite;
 import org.junit.platform.suite.api.SelectClasses;
 import org.junit.platform.suite.api.Suite;
 import suite.basic.BasicTest;
-import suite.operator.V2ApiOperatorLocalOnlyTest;
 import suite.operator.V2ApiOperatorTest;
 
 @Suite
